random(patch): Fix format specifiers and type cast variadic arguments to double to mitigate implicit conversions

+++ examples/random.c
@@ -13,22 +13,22 @@
 
 void initialize(void) {
     long z = time(NULL); // use current time as seed
-    printf("initialize: %d\n", z);
+    printf("initialize: %ld\n", z);
     srand(z); // seed the rng using current time
     printf("srand: seeded\n");
 }
 
 float normalize(void) {
     long z = rand(); // return a number between z and RAND_MAX
-    printf("rand: %d\n", z);
+    printf("rand: %ld\n", z);
     float norm = (float) z / RAND_MAX;
-    printf("normalize: %f\n", norm);
+    printf("normalize: %f\n", (double) norm);
     return norm;
 }
 
 void interval(float n) {
     float i = n * 2 - 1; // bind to interval [-1, 1]
-    printf("interval: %f\n", i);
+    printf("interval: %f\n", (double) i);
 }
 
 int main(void) {
